refactor(capture-sdk): Apply PR suggestions
- change package name to conform to Google's recomnendation
- change version name from relative to fix version for Mixpanel
- remove new permissions

PP-438
