+++ capture-sdk/sdk/src/main/AndroidManifest.xml
@@ -3,8 +3,6 @@
     package="net.gini.android.capture">
 
     <uses-permission android:name="android.permission.VIBRATE" />
-    <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.READ_PHONE_STATE" />
 
     <application>
         <activity
